+++ test/ModuleLoader.spec.js
@@ -22,7 +22,6 @@ describe( 'ModuleLoader', function() {
 		it( 'should not allow modules with empty or null names', function() {
 			expect( () => moduleLoader.register( null ) ).to.throw( Error );
 			expect( () => moduleLoader.register( undefined ) ).to.throw( Error );
-			expect( () => moduleLoader.register( null ) ).to.throw( Error );
 			expect( () => moduleLoader.register( '' ) ).to.throw( Error );
 			expect( () => moduleLoader.register( 2 ) ).to.throw( Error );
 		} );
@@ -126,7 +125,7 @@ describe( 'ModuleLoader', function() {
 			expect( () => moduleLoader.register( [ _.noop, _.noop ] ) ).to.not.throw( Error );
 		} );
 
-		it( 'Should throw an exception if already started', function() {
+		it( 'should throw an exception if already started', function() {
 			expect( () => moduleLoader.register( 'a' ) ).to.not.throw( Error );
 			expect( () => moduleLoader.register( { name: 'b', dependencies: 'a' } ) ).to.not.throw( Error );
 			return moduleLoader.start().then( () => {
@@ -152,7 +151,7 @@ describe( 'ModuleLoader', function() {
 	describe( '#start', function() {
 
 		it( 'should do nothing if no module is registered', function() {
-			expect( () => moduleLoader.start() ).to.not.throw();
+			return expect( moduleLoader.start() ).to.be.eventually.fulfilled;
 		} );
 
 		it( 'should run modules that have no dependencies', function() {
@@ -173,33 +172,35 @@ describe( 'ModuleLoader', function() {
 
 		} );
 
-		it( 'should throw an error if a circular dependency is found', function() {
+		it( 'should eventually throw an error if a circular dependency is found', function() {
 
 			moduleLoader.register( { name: 'a', dependencies: [] } );
 			moduleLoader.register( { name: 'b', dependencies: [ 'c' ] } );
 			moduleLoader.register( { name: 'c', dependencies: [ 'a', 'b' ] } );
-			expect( () => moduleLoader.start() ).to.throw( Error );
+			return expect( moduleLoader.start() ).to.eventually.be.rejected;
 
 		} );
 
-		// Please note that this test is actually included in the circular dependency one since,
-		// without a root module, a circular dependency is mathematically required to happen.
-		// The test is only here to cover for the more descriptive error message returned to
-		// the user in this particular case.
-		it( 'should throw an error if no root modules are found', function() {
+		/**
+		 * Please note that this test is actually included in the circular dependency one since,
+		 * without a root module, a circular dependency is mathematically required to happen.
+		 * The test is only here to cover for the more descriptive error message returned to
+		 * the user in this particular case.
+		 */
+		it( 'should eventually throw an error if no root modules are found', function() {
 
 			moduleLoader.register( { name: 'a', dependencies: [ 'b' ] } );
 			moduleLoader.register( { name: 'b', dependencies: [ 'c' ] } );
 			moduleLoader.register( { name: 'c', dependencies: [ 'a' ] } );
-			expect( () => moduleLoader.start() ).to.throw( Error );
+			return expect( moduleLoader.start() ).to.eventually.be.rejected;
 
 		} );
 
-		it( 'should throw an error if an unknown dependency is found', function() {
+		it( 'should eventually throw an error if an unknown dependency is found', function() {
 
 			moduleLoader.register( { name: 'a', dependencies: [] } );
 			moduleLoader.register( { name: 'b', dependencies: [ 'c' ] } );
-			expect( () => moduleLoader.start() ).to.throw( Error );
+			return expect( moduleLoader.start() ).to.eventually.be.rejected;
 
 		} );
 
@@ -235,19 +236,14 @@ describe( 'ModuleLoader', function() {
 
 		} );
 
-		it( 'should return a resolved Promise when no modules are registered', function() {
-			expect( moduleLoader.start() ).to.be.an.instanceOf( Bluebird );
-			return expect( moduleLoader.start() ).to.eventually.be.fulfilled;
-		} );
-
-		it( 'should return a resolved Promise when some modules are registered', function() {
+		it( 'should eventually fullfill a Promise when some modules are registered', function() {
 			moduleLoader.register( { name: 'a', dependencies: [] } );
 			moduleLoader.register( { name: 'b', dependencies: [ 'a' ] } );
 			expect( moduleLoader.start() ).to.be.an.instanceOf( Bluebird );
 			return expect( moduleLoader.start() ).to.eventually.be.fulfilled;
 		} );
 
-		it( 'should return a rejected Promise if a start function does not return a value', function() {
+		it( 'should eventually reject a Promise if a start function does not return a value', function() {
 			moduleLoader.register( { name: 'a', dependencies: [], start: function() { } } );
 			expect( moduleLoader.start() ).to.be.an.instanceOf( Bluebird );
 			return expect( moduleLoader.start() ).to.eventually.be.rejected;
@@ -294,6 +290,11 @@ describe( 'ModuleLoader', function() {
 			return moduleLoader.start();
 		} );
 
+		/**
+		 * A module value changed by another module is dangerous, as it is prone to race conditions and order requirements which cannot be satisfied by this library.
+		 * On the other hand, forcing a different instance of a dependency per module means that no communication can ever be accomplished between modules.
+		 * Between the two evils, we prefer the shared instance way, moving the race condition problems up in the hands of the user.
+		 */
 		it( 'should allow different modules to share state between dependencies', function() {
 			moduleLoader.register( { name: 'a', dependencies: [], start: () => { return { value: 1 }; } } );
 			moduleLoader.register( {
@@ -306,6 +307,7 @@ describe( 'ModuleLoader', function() {
 			moduleLoader.register( {
 				name: 'c', dependencies: [ 'a', 'b' ], start: ( a, b ) => {
 					expect( a.value ).to.be.eql( 0 );
+					expect( b ).to.be.eql( 1 );
 					return 1;
 				}
 			} );
@@ -368,6 +370,30 @@ describe( 'ModuleLoader', function() {
 			} );
 		} );
 
+		it( 'should eventually reject undefined return values for named modules', function() {
+			moduleLoader.register( {
+				name: 'a',
+				start: () => undefined
+			} );
+			return expect( moduleLoader.start() ).to.be.eventually.rejected;
+		} );
+
+		it( 'should eventually allow undefined return values for anonymous modules', function() {
+			moduleLoader.register( {
+				start: () => undefined
+			} );
+			return expect( moduleLoader.start() ).to.be.eventually.fulfilled;
+		} );
+
+		it( 'should differentiate between anonymous modules', function() {
+			let counter = 0;
+			let startA = () => counter += 5;
+			let startB = () => counter -= 3;
+			moduleLoader.register( { start: startA } );
+			moduleLoader.register( { start: startB } );
+			return moduleLoader.start().then( () => expect( counter ).to.be.eql( 2 ) );
+		} );
+
 	} );
 
 	describe( '#resolve', function() {
@@ -391,11 +417,14 @@ describe( 'ModuleLoader', function() {
 		} );
 
 		it( 'should throw an error if the given argument is not a valid dependency name', function() {
-			return Promise.all( [
-				expect( moduleLoader.resolve() ).to.be.eventually.rejected,
-				expect( moduleLoader.resolve( 2 ) ).to.be.eventually.rejected,
-				expect( moduleLoader.resolve( [ 2 ] ) ).to.be.eventually.rejected
-			] );
+			expect( () => moduleLoader.resolve() ).to.throw();
+			expect( () => moduleLoader.resolve( 2 ) ).to.throw();
+			expect( () => moduleLoader.resolve( [ 'a', 2 ] ) ).to.throw();
+			expect( () => moduleLoader.resolve( '$' ) ).to.throw();
+			expect( () => moduleLoader.resolve( [ 'a', '$' ] ) ).to.throw();
+			expect( () => moduleLoader.resolve( null ) ).to.throw();
+			expect( () => moduleLoader.resolve( undefined ) ).to.throw();
+			expect( () => moduleLoader.resolve( {} ) ).to.throw();
 		} );
 
 		it( 'should return undefined if the given argument is not a registered dependency', function() {
@@ -422,11 +451,11 @@ describe( 'ModuleLoader', function() {
 			return expect( moduleLoader.resolve( 'c' ) ).to.be.eventually.eql( 4 );
 		} );
 
-		it( 'should allow resolution of multiple modules', function() {
+		it( 'should eventually allow resolution of multiple modules', function() {
 			return expect( moduleLoader.resolve( [ 'a', 'b' ] ) ).to.be.eventually.deep.equal( [ 1, 2 ] );
 		} );
 
-		it( 'should allow resolution of all listed modules', function() {
+		it( 'should eventually allow resolution of all listed modules', function() {
 			let resolution = moduleLoader.resolve( moduleLoader.list() );
 			return expect( resolution ).to.be.eventually.fulfilled;
 		} );
@@ -646,29 +675,28 @@ describe( 'ModuleLoader', function() {
 
 		it( 'should register dependencies for injection', function() {
 			[ 'a', 'a2', 'b', 'c', 'd' ].forEach( regFile );
-			moduleLoader.start();
-
-			return Promise.all( [
-				expect( moduleLoader.resolve( 'a' ) ).to.eventually.have.property( 'ok', 1 ),
-				expect( moduleLoader.resolve( 'b' ) ).to.eventually.have.property( 'ok', 2 ),
-				expect( moduleLoader.resolve( 'c' ) ).to.eventually.have.property( 'ok', 3 ),
-				expect( moduleLoader.resolve( 'd' ) ).to.eventually.have.property( 'ok', 4 ),
-				expect( moduleLoader.resolve( 'a2' ) ).to.eventually.have.property( 'ok', 5 )
-			] );
-
+			return moduleLoader.start().then( () =>
+				Promise.all( [
+					expect( moduleLoader.resolve( 'a' ) ).to.eventually.have.property( 'ok', 1 ),
+					expect( moduleLoader.resolve( 'b' ) ).to.eventually.have.property( 'ok', 2 ),
+					expect( moduleLoader.resolve( 'c' ) ).to.eventually.have.property( 'ok', 3 ),
+					expect( moduleLoader.resolve( 'd' ) ).to.eventually.have.property( 'ok', 4 ),
+					expect( moduleLoader.resolve( 'a2' ) ).to.eventually.have.property( 'ok', 5 )
+				] )
+			);
 		} );
 
 		it( 'should register dependencies with camel-case naming', function() {
 			[ 'a', 'ClassName', 'long-name' ].forEach( regFile );
-			moduleLoader.start();
-
-			return Promise.all( [
-				expect( moduleLoader.resolve( 'a' ) ).to.not.be.eventually.undefined,
-				expect( moduleLoader.resolve( 'className' ) ).to.not.be.eventually.undefined,
-				expect( moduleLoader.resolve( 'longName' ) ).to.not.be.eventually.undefined,
-				expect( moduleLoader.resolve( 'ClassName' ) ).to.be.eventually.undefined,
-				expect( moduleLoader.resolve( 'long-name' ) ).to.be.eventually.undefined
-			] );
+			return moduleLoader.start().then( () =>
+				Promise.all( [
+					expect( moduleLoader.resolve( 'a' ) ).to.not.be.eventually.undefined,
+					expect( moduleLoader.resolve( 'className' ) ).to.not.be.eventually.undefined,
+					expect( moduleLoader.resolve( 'longName' ) ).to.not.be.eventually.undefined,
+					expect( moduleLoader.resolve( 'ClassName' ) ).to.be.eventually.undefined,
+					expect( moduleLoader.resolve( 'long-name' ) ).to.be.eventually.undefined
+				] )
+			);
 		} );
 
 	} );
@@ -684,53 +712,50 @@ describe( 'ModuleLoader', function() {
 		} );
 
 		it( 'should register all modules in a folder', function() {
-
 			moduleLoader.registerDirectory( testDirectory );
-			moduleLoader.start();
-
-			return Promise.all( [
-				expect( moduleLoader.resolve( 'a' ) ).to.not.be.eventually.undefined,
-				expect( moduleLoader.resolve( 'database' ) ).to.not.be.eventually.undefined,
-				expect( moduleLoader.resolve( 'webServer' ) ).to.not.be.eventually.undefined,
-				expect( moduleLoader.resolve( 'logger' ) ).to.be.eventually.undefined
-			] );
-
+			return moduleLoader.start().then( () =>
+				Promise.all( [
+					expect( moduleLoader.resolve( 'a' ) ).to.not.be.eventually.undefined,
+					expect( moduleLoader.resolve( 'database' ) ).to.not.be.eventually.undefined,
+					expect( moduleLoader.resolve( 'webServer' ) ).to.not.be.eventually.undefined,
+					expect( moduleLoader.resolve( 'logger' ) ).to.be.eventually.undefined
+				] )
+			);
 		} ).slow( 200 );
 
 		it( 'should register all modules in a folder and its subfolders, if asked', function() {
-
 			moduleLoader.registerDirectory( testDirectory, true );
-			moduleLoader.start();
-
-			return Promise.all( [
-				expect( moduleLoader.resolve( 'a' ) ).to.not.be.eventually.undefined,
-				expect( moduleLoader.resolve( 'database' ) ).to.not.be.eventually.undefined,
-				expect( moduleLoader.resolve( 'webServer' ) ).to.not.be.eventually.undefined,
-				expect( moduleLoader.resolve( 'logger' ) ).to.not.be.eventually.undefined
-			] );
-
+			return moduleLoader.start().then( () =>
+				Promise.all( [
+					expect( moduleLoader.resolve( 'a' ) ).to.not.be.eventually.undefined,
+					expect( moduleLoader.resolve( 'database' ) ).to.not.be.eventually.undefined,
+					expect( moduleLoader.resolve( 'webServer' ) ).to.not.be.eventually.undefined,
+					expect( moduleLoader.resolve( 'logger' ) ).to.not.be.eventually.undefined
+				] )
+			);
 		} ).slow( 200 );
 
 	} );
 
 	describe( '#registerValue', function() {
 
-		it( 'should resolve the registered value', function() {
+		it( 'should eventually resolve the registered value', function() {
 			moduleLoader.registerValue( 'a', 1 );
 			moduleLoader.registerValue( 'b', 'b' );
 			moduleLoader.registerValue( 'c', [ 1, 2 ] );
 			moduleLoader.registerValue( 'd', { d: 1 } );
-			moduleLoader.start();
-			return Promise.all( [
-				expect( moduleLoader.resolve( 'a' ) ).to.be.eventually.deep.equal( 1 ),
-				expect( moduleLoader.resolve( 'b' ) ).to.be.eventually.deep.equal( 'b' ),
-				expect( moduleLoader.resolve( 'c' ) ).to.be.eventually.deep.equal( [ 1, 2 ] ),
-				expect( moduleLoader.resolve( 'd' ) ).to.be.eventually.deep.equal( { d: 1 } )
-			] );
+			return moduleLoader.start().then( () =>
+				Promise.all( [
+					expect( moduleLoader.resolve( 'a' ) ).to.be.eventually.deep.equal( 1 ),
+					expect( moduleLoader.resolve( 'b' ) ).to.be.eventually.deep.equal( 'b' ),
+					expect( moduleLoader.resolve( 'c' ) ).to.be.eventually.deep.equal( [ 1, 2 ] ),
+					expect( moduleLoader.resolve( 'd' ) ).to.be.eventually.deep.equal( { d: 1 } )
+				] )
+			);
 		} );
 
 		it( 'should not allow null or undefined values', function() {
-			expect( () => moduleLoader.registerValue( 'e', undefined ) ).to.throw( Error );
+			expect( () => moduleLoader.registerValue( 'e', null ) ).to.throw( Error );
 			expect( () => moduleLoader.registerValue( 'e', undefined ) ).to.throw( Error );
 		} );